+++ tests/exports.spec.ts
@@ -1,0 +1,12 @@
+import * as bundle from '../src';
+import { HomeAssistantStylesManager } from '../src';
+import { getCSSRulesString } from '../src/utilities';
+
+describe('public exports', () => {
+    it('should export the proper elements', () => {
+        expect(bundle).toEqual({
+            HomeAssistantStylesManager,
+            getCSSRulesString
+        });
+    });
+});